README:
Following the previous commit and the advices from Microsoft,
I have provided a complete section for Microsoft Windows.
Now, there are 2 build options: along with php source code or
standalone with with the PHP develpment environment.

You should read:
  https://github.com/Microsoft/php-sdk-binary-tools/pull/48#issuecomment-453749874

if you want to understand the rationals.

Down the road, I hope it can become a reference for:
   https://wiki.php.net/internals/windows

TODO: extend this tutorial and illustrate it using appveyor CI.
XXX: why are bison, sed, re2c required knowing they are not used ?!